Added experimental support for enumerating subproperties.
Subproperties are a way to index e.g. different features of part of
speech in a single field. See DocIndexerOpenSonar for an example. Now,
if you've indexed subproperties, you can enumerate them in BlackLab
Server by going to /bls/indexName/fields/contents?subprops=pos to view
subproperties in the "pos" property of the "contents" field.

+++ core/src/main/java/nl/inl/blacklab/forwardindex/TermsImplV3.java
@@ -359,7 +359,7 @@ class TermsImplV3 extends Terms {
 					if (!useBlockBasedTermsFile) {
 						long fileLength = 2 * BYTES_PER_INT + (n + 1) * BYTES_PER_INT + termStringsByteSize + NUM_SORT_BUFFERS * BYTES_PER_INT * n;
 						fc.truncate(fileLength); // truncate if necessary
-						buf = fc.map(MapMode.READ_WRITE, (long) 0, fileLength);
+						buf = fc.map(MapMode.READ_WRITE, 0, fileLength);
 						buf.putInt(n); // Start with the number of terms
 						ib = buf.asIntBuffer();
